add menu to run playbook in yaml file, refine commands layout
