+++ src/extension.ts
@@ -5,11 +5,13 @@ import * as vscode from 'vscode';
 import * as utilities from './utilities';
 import { AzureAccount } from './azure-account.api';
 import { Constants } from './constants';
-import { TerminalRunner } from './terminalRunner';
 import { CloudShellRunner } from './cloudShellRunner';
 import { TerminalExecutor } from './terminalExecutor';
 import { AnsibleCompletionItemProvider } from './ansibleCompletionItemProvider';
 import { TelemetryClient } from './telemetryClient';
+import { Option } from './baseRunner';
+import { DockerRunner } from './dockerRunner';
+import { LocalAnsibleRunner } from './localAnsibleRunner';
 
 export function activate(context: vscode.ExtensionContext) {
     console.log('Congratulations, your extension "vscode-ansible" is now active!');
@@ -22,15 +24,20 @@ export function activate(context: vscode.ExtensionContext) {
 
     utilities.generateCredentialsFile();
 
-    var terminalRunner = new TerminalRunner(outputChannel);
+    var dockerRunner = new DockerRunner(outputChannel);
+    var localansibleRunner = new LocalAnsibleRunner(outputChannel);
     var cloudShellRunner = new CloudShellRunner(outputChannel);
 
-    context.subscriptions.push(vscode.commands.registerCommand('vscode-ansible.playbook-in-terminal', () => {
-        terminalRunner.runPlaybook();
+    context.subscriptions.push(vscode.commands.registerCommand('vscode-ansible.playbook-in-docker', (playbook) => {
+        dockerRunner.runPlaybook(playbook ? playbook.fsPath : null);
     }));
 
-    context.subscriptions.push(vscode.commands.registerCommand('vscode-ansible.cloudshell', () => {
-        cloudShellRunner.runPlaybook();
+    context.subscriptions.push(vscode.commands.registerCommand('vscode-ansible.playbook-in-localansible', (playbook) => {
+        localansibleRunner.runPlaybook(playbook ? playbook.fsPath : null);
+    }));
+
+    context.subscriptions.push(vscode.commands.registerCommand('vscode-ansible.cloudshell', (playbook) => {
+        cloudShellRunner.runPlaybook(playbook ? playbook.fsPath : null);
     }));
 
     context.subscriptions.push(vscode.window.onDidCloseTerminal((closedTerminal: vscode.Terminal) => {
